requirements.txt
numpy
scikit-learn
spacy
nltk
joblib
fastapi
uvicorn[standard]
pymongo
pydantic-settings
google-api-python-client
google-auth
requests
lime
shap
langdetect

